Refactor core test putting model tests in own file.
